refactor(batch): separate concurrency and file-size policies
