+++ __tests__/batch-lint.spec.ts
@@ -3,14 +3,9 @@ import { availableParallelism, tmpdir } from "os";
 import * as path from "path";
 import { Piscina } from "piscina";
 import type { LintMdRulesConfig } from "@lint-md/core";
-import {
-  STAT_CONCURRENCY_LIMIT,
-  batchLint,
-  getMaxFileSize,
-  keepLintItem,
-  resolveAdaptiveConcurrency,
-  runTasksWithLimit,
-} from "../src/utils/batch-lint";
+import { resolveAdaptiveConcurrency } from "../src/utils/adaptive-concurrency";
+import { batchLint, keepLintItem } from "../src/utils/batch-lint";
+import { STAT_CONCURRENCY_LIMIT, getMaxFileSize } from "../src/utils/file-stat";
 import type { BatchLintItem } from "../src/types";
 import { makeNotAppliedFix } from "./helpers/not-applied-fix";
 
@@ -25,41 +20,6 @@ const RULES_NO_EMPTY_LIST: LintMdRulesConfig = {
 };
 
 const TRIGGER_CONTENT = "1. hello\n2.\n";
-
-describe("runTasksWithLimit", () => {
-  test("respects concurrency limit", async () => {
-    let running = 0;
-    let maxRunning = 0;
-
-    const tasks = Array.from({ length: 10 }, () => async () => {
-      running++;
-      maxRunning = Math.max(maxRunning, running);
-      await new Promise((resolve) => setTimeout(resolve, 10));
-      running--;
-      return true;
-    });
-
-    await runTasksWithLimit(tasks, 2);
-    expect(maxRunning).toBe(2);
-  });
-
-  test("preserves result order", async () => {
-    const tasks = [3, 1, 4, 1, 5].map((n) => async () => n);
-    const result = await runTasksWithLimit(tasks, 2);
-    expect(result).toEqual([3, 1, 4, 1, 5]);
-  });
-
-  test("handles empty task list", async () => {
-    const result = await runTasksWithLimit([], 3);
-    expect(result).toEqual([]);
-  });
-
-  test("limit greater than tasks count still works", async () => {
-    const tasks = [1, 2, 3].map((n) => async () => n * 10);
-    const result = await runTasksWithLimit(tasks, 10);
-    expect(result).toEqual([10, 20, 30]);
-  });
-});
 
 describe("batchLint", () => {
   let tmpDir: string;
