+++ src/utils/batch-lint.ts
@@ -1,40 +1,10 @@
 import path from "path";
 import { existsSync } from "fs";
-import { stat } from "fs/promises";
-import { availableParallelism } from "os";
 import { Piscina } from "piscina";
 import type { LintMdRulesConfig } from "@lint-md/core";
-import type { BatchLintItem, LintWorkerOptions, ThreadCount } from "../types";
+import type { BatchLintItem, LintWorkerOptions } from "../types";
 import { isIncompleteFix } from "./report-incomplete-fixes";
-
-const ONE_MIB = 1024 * 1024;
-const FIVE_MIB = 5 * ONE_MIB;
-const ADAPTIVE_MEDIUM_CAP = 2;
-const ADAPTIVE_LARGE_FILE_THRESHOLD = ONE_MIB;
-const ADAPTIVE_HUGE_FILE_THRESHOLD = FIVE_MIB;
-// Upper bound on concurrent stat() fds in getMaxFileSize (see #80).
-export const STAT_CONCURRENCY_LIMIT = 128;
-
-export async function runTasksWithLimit<T>(
-  tasks: (() => Promise<T>)[],
-  limit: number
-): Promise<T[]> {
-  const results: T[] = [];
-  let index = 0;
-
-  async function runNext(): Promise<void> {
-    while (index < tasks.length) {
-      const i = index++;
-      results[i] = await tasks[i]();
-    }
-  }
-
-  const workers = Array.from({ length: Math.min(limit, tasks.length) }, () =>
-    runNext()
-  );
-  await Promise.all(workers);
-  return results;
-}
+import { runTasksWithLimit } from "./run-tasks-with-limit";
 
 const resolveWorkerFilename = (): string => {
   const compiled = path.resolve(__dirname, "./lint-worker.js");
@@ -42,72 +12,6 @@ const resolveWorkerFilename = (): string => {
     return compiled;
   }
   return path.resolve(__dirname, "../../lib/src/utils/lint-worker.js");
-};
-
-// getMaxFileSize() stats every file but bounds the in-flight stat calls to
-// STAT_CONCURRENCY_LIMIT via runTasksWithLimit, avoiding an N-fd filesystem
-// burst on very large repositories (see #80).
-// We bound concurrency rather than short-circuit on a >= 5 MiB file: this
-// function must return the TRUE maximum size, because src/lint-md.ts logs
-// `max file ${maxMiB} MiB` under --threads auto --dev. An early return would
-// make that log (and any future consumer) inaccurate. STAT_CONCURRENCY_LIMIT
-// is an empirical cap (the issue suggests 64/128), not benchmark-derived.
-export const getMaxFileSize = async (filePaths: string[]): Promise<number> => {
-  if (filePaths.length === 0) {
-    return 0;
-  }
-  const sizes = await runTasksWithLimit(
-    filePaths.map(
-      (filePath) => () => stat(filePath).then((stats) => stats.size)
-    ),
-    STAT_CONCURRENCY_LIMIT
-  );
-  return sizes.reduce((max, current) => (current > max ? current : max), 0);
-};
-
-export interface AdaptiveConcurrencyDecision {
-  concurrency: number;
-  maxFileSize: number | null;
-  requestedConcurrency: number;
-}
-
-export const resolveAdaptiveConcurrency = async (
-  threadCount: ThreadCount,
-  mdFilePaths: string[]
-): Promise<AdaptiveConcurrencyDecision> => {
-  const requestedConcurrency =
-    typeof threadCount === "number" ? threadCount : availableParallelism();
-
-  if (mdFilePaths.length === 0) {
-    return {
-      concurrency: 0,
-      maxFileSize: threadCount === "auto" ? 0 : null,
-      requestedConcurrency,
-    };
-  }
-
-  if (typeof threadCount === "number") {
-    return {
-      concurrency: Math.min(Math.max(threadCount, 1), mdFilePaths.length),
-      maxFileSize: null,
-      requestedConcurrency,
-    };
-  }
-
-  const maxFileSize = await getMaxFileSize(mdFilePaths);
-
-  let limit = requestedConcurrency;
-  if (maxFileSize >= ADAPTIVE_HUGE_FILE_THRESHOLD) {
-    limit = 1;
-  } else if (maxFileSize >= ADAPTIVE_LARGE_FILE_THRESHOLD) {
-    limit = Math.min(limit, ADAPTIVE_MEDIUM_CAP);
-  }
-
-  return {
-    concurrency: Math.min(Math.max(limit, 1), mdFilePaths.length),
-    maxFileSize,
-    requestedConcurrency,
-  };
 };
 
 // Keep a file's result when it has lint findings, or — in fix mode — when
