+++ src/utils/run-tasks-with-limit.ts
@@ -1,0 +1,20 @@
+export async function runTasksWithLimit<T>(
+  tasks: (() => Promise<T>)[],
+  limit: number
+): Promise<T[]> {
+  const results: T[] = [];
+  let index = 0;
+
+  async function runNext(): Promise<void> {
+    while (index < tasks.length) {
+      const currentIndex = index++;
+      results[currentIndex] = await tasks[currentIndex]();
+    }
+  }
+
+  const workers = Array.from({ length: Math.min(limit, tasks.length) }, () =>
+    runNext()
+  );
+  await Promise.all(workers);
+  return results;
+}
